arm64: hibernate: idmap the single page that holds the copy page routines
To resume from hibernate, the contents of memory are restored from
the swap image. This may overwrite any page, including the running
kernel and its page tables.

Hibernate uses a temporary set of page tables in TTBR1, but to
break-before-make this mapping, its restore code must run from TTBR0.
Hibernate copies the code it uses to do the restore into a single
page that it knows won't be overwritten, and maps it with page tables
built from pages that won't be overwritten. This is loaded in TTBR0.

Today the address it uses for this mapping is arbitrary, but to allow
kexec to reuse this code, it needs to be idmapped, as kexec additionally
needs to disable the MMU. To idmap the page we must avoid the kernel
helpers that have VA_BITS baked in.

Convert trans_pgd_map_page() to take a single PA, and idmap it.
The page tables are built in the reverse order to normal using
pfn_pte() to stir in any bits between 52:48. T0SZ is always increased
to cover 48bits, or 52 if the copy code has bits 52:48 in its PA.

Extending this idmap to 48 bits ensures the 'extended idmap' behaviour
can be tested on any machine by selecting a lower VA_BITS.

Signed-off-by: James Morse <james.morse@arm.com>