arm64: kexec: Enter the kdump kernel earlier
The kdump kernel is always loaded into its own carveout of memory.
Because we now do the cache-maintenance earlier, there is very little
for machine_kexec() and its relocation code to do.

Spot the in-place kdump image, and start it directly. This bypasses
the relocation code (which does nothing), and uses the hyp-stub API
if we booted at EL2.

This also catches kexec it managed to be placed such that no
relocation was needed.

This allows us to do the relocation specific things after kdump
has left this world. Using the hyp-stub API will let us change
cpu_soft_restart() to not change exception level.

Signed-off-by: James Morse <james.morse@arm.com>