arm64: hibernate: Use the hyp stub instead of re-inventing it
Hibernate contains a set of temporary EL2 vectors used to 'park'
EL2 somewhere safe while all the memory is thrown in the air.

All this really does is HVC_SET_VECTORS. Now that the hyp-stub
is self contained, copy it somewhere that won't be overwritten,
and use that.

This saves having 2K of duplicate vector, that might not get used.

Signed-off-by: James Morse <james.morse@arm.com>