arm64: kexec: Copy and install a copy of the hyp-stub
Before we can do the kexec relocation with the MMU enable, we need to
move the relocation code to EL1. The payload may need to be called
at EL2, so we need a way to call code at EL2. This is what the hyp-stub's
HVC_SOFT_RESTART does.

Create a copy of the hyp-stub in a page that won't be overwritten during
relocation, and install it before we start.

Signed-off-by: James Morse <james.morse@arm.com>