arm64: kexec: Read relocation parameters from struct kimage
Arguments to the kexec relocation code are about to become precious.
The hyp-stub API limits the number of arguments we can pass to the kexec
relocation code, as it must also be able to pass HVC_SOFT_RESTART.

Instead of expanding the two implementations, make the caller smarter.

The three values we use in the relocation code are all read from
the same struct. Pass the physical address of the struct.

Now that we read struct kimage from the relocation code, we must
also ensure it is clean to the PoC.

Signed-off-by: James Morse <james.morse@arm.com>