arm64: kexec: Move PoC maintenance to a single place
Now that we have machine_kexec_post_load(), we can do the cache
maintenance in a single place, instead of ad-hoc when we get the
chance.

Kdump used to do this when the crashkernel region was unmapped
by the protect call. Regular kexec cleans the relocation list
during machine_kexec(), and maybe the segments if its empty,
while being careful not to clean the kdump kernel as it won't be
mapped.

This is going to be simpler if its done in machine_kexec_post_load()
as the crashkernel region has't yet been unmapped.

Signed-off-by: James Morse <james.morse@arm.com>