arm64: kexec: Move size calculation to C
To disable the EL1 MMU after relocation, the body of _cpu_soft_restart
needs to move into the kexec relocation code. As kexec is its only caller
this should be straight forward.

But until we idmap the kexec relocation code, the relocation code would
need to be in the idmap. Putting kexec's relocation code in the idmap
confuses the size calculation.

Update the asm annotations on the kexec relocation code, and move the
size calculation to C.

Signed-off-by: James Morse <james.morse@arm.com>