arm64: kexec: merge cpu_soft_restart() with its caller
cpu_soft_restart() disables the EL1 MMU and calls the provided address.

Once we run kexec with the MMU enabled, we will need to do this disable
after relocation in code that is copied. We also need to be able to call
this for kdump when no relocations are needed.

This means doing slightly different things with the two callers of
__cpu_soft_restart(). Merge the function in cpu-reset.h into its two
callers.

Signed-off-by: James Morse <james.morse@arm.com>