arm64: kexec: Pass the phy_to_virt offset to the relocation code
The kexec list of source/destination pages is physically addressed.
Once we leave the MMU enabled, only the linear-map will be accesible.

Re-purpose one of the zeros passed to the relocation code as the
phys_to_virt offset. The relocation code can add this to any
physical address to map it into the linear region.

Signed-off-by: James Morse <james.morse@arm.com>