vfio_pci: save and restore vfio_pci_device "keepalive" state
The PCI passthrough device is managed by vfio_pci driver. When the
device is set to "keepalive" state, vfio_pci driver has its own
private state that can not be restored without writing to hardware.
Because of this, when vfio_pci instance is released, the driver
needs to save its private software keepalive state, so that when
new instance is opened, it can restore the instance state with
saved keepalive state.

Signed-off-by: Jason Zeng <jason.zeng@intel.com>